refactor: move owner register api to owner controller
