+++ src/2-Services/Monolith/Api/Monolith.Api/Controllers/OwnerApiController.cs
@@ -8,6 +8,10 @@ using TaskoMask.BuildingBlocks.Web.ApiContracts;
 using TaskoMask.BuildingBlocks.Contracts.Dtos.Workspace.Owners;
 using TaskoMask.BuildingBlocks.Domain.Services;
 using TaskoMask.BuildingBlocks.Contracts.Services;
+using TaskoMask.BuildingBlocks.Contracts.Enums;
+using TaskoMask.BuildingBlocks.Contracts.Dtos.Authorization.Users;
+using TaskoMask.BuildingBlocks.Web.MVC.Services.Authentication.JwtAuthentication;
+using TaskoMask.BuildingBlocks.Contracts.Models;
 
 namespace TaskoMask.Services.Monolith.Api.Controllers
 {
@@ -17,19 +21,49 @@ namespace TaskoMask.Services.Monolith.Api.Controllers
         #region Fields
 
         private readonly IOwnerService _ownerService;
+        private readonly IJwtAuthenticationService _jwtAuthenticationService;
 
         #endregion
 
         #region Ctors
 
-        public OwnerApiController(IOwnerService ownerService, IAuthenticatedUserService authenticatedUserService) : base(authenticatedUserService)
+        public OwnerApiController(IOwnerService ownerService, IAuthenticatedUserService authenticatedUserService, IJwtAuthenticationService jwtAuthenticationService) : base(authenticatedUserService)
         {
             _ownerService = ownerService;
+            _jwtAuthenticationService = jwtAuthenticationService;
         }
 
         #endregion
 
         #region Public Methods
+
+
+
+        /// <summary>
+        /// register new owner - return jwt token if register is success
+        /// </summary>
+        [HttpPost]
+        [Route("owner")]
+        [AllowAnonymous]
+        public async Task<Result<UserJwtTokenDto>> Register([FromBody] RegisterOwnerDto input)
+        {
+            //TODO refactor with Identity Server
+
+
+            //create owner with default workspace
+            var createCommandResult = await _ownerService.RegisterAndSeedDefaultWorkspaceAsync(input);
+            if (!createCommandResult.IsSuccess)
+                return Result.Failure<UserJwtTokenDto>(createCommandResult.Errors, createCommandResult.Message);
+
+            var user = GetAuthenticatedUserModel(createCommandResult.Value.EntityId, input);
+
+            //generate jwt token
+            var token = _jwtAuthenticationService.GenerateJwtToken(user);
+
+            return Result.Success(value: new UserJwtTokenDto { JwtToken = token });
+        }
+
+
 
 
 
@@ -58,6 +92,27 @@ namespace TaskoMask.Services.Monolith.Api.Controllers
             return await _ownerService.UpdateProfileAsync(input);
         }
 
+
+
+        #endregion
+
+
+        #region Private Methods
+
+
+        /// <summary>
+        /// 
+        /// </summary>
+        private AuthenticatedUserModel GetAuthenticatedUserModel(string id, RegisterOwnerDto owner)
+        {
+            return new AuthenticatedUserModel
+            {
+                Id = id,
+                DisplayName = owner.DisplayName,
+                Email = owner.Email,
+                UserName = owner.Email,
+            };
+        }
 
 
         #endregion
